Make Integer and Float wrappers transparent
- Annotate these types with `#[repr(transparent)]`.
- Implement `fmt::Debug` as a passthrough to the inner primitive type.
- Implement `fmt::Display` as a passthrough to the inner primitive type.
- Move `float::init` to a `float::mruby` module.
- Remove `artichoke_` prefix from `Integer` C mruby trampolines.
