+++ artichoke-backend/src/extn/core/integer/mruby.rs
@@ -12,13 +12,13 @@ pub fn init(interp: &mut Artichoke) -> InitializeResult<()> {
     }
     let spec = class::Spec::new("Integer", INTEGER_CSTR, None, None)?;
     class::Builder::for_spec(interp, &spec)
-        .add_method("chr", artichoke_integer_chr, sys::mrb_args_opt(1))?
-        .add_method("[]", artichoke_integer_element_reference, sys::mrb_args_req(1))?
-        .add_method("/", artichoke_integer_div, sys::mrb_args_req(1))?
-        .add_method("allbits?", artichoke_integer_is_allbits, sys::mrb_args_req(1))?
-        .add_method("anybits?", artichoke_integer_is_anybits, sys::mrb_args_req(1))?
-        .add_method("nobits?", artichoke_integer_is_nobits, sys::mrb_args_req(1))?
-        .add_method("size", artichoke_integer_size, sys::mrb_args_none())?
+        .add_method("chr", integer_chr, sys::mrb_args_opt(1))?
+        .add_method("[]", integer_element_reference, sys::mrb_args_req(1))?
+        .add_method("/", integer_div, sys::mrb_args_req(1))?
+        .add_method("allbits?", integer_is_allbits, sys::mrb_args_req(1))?
+        .add_method("anybits?", integer_is_anybits, sys::mrb_args_req(1))?
+        .add_method("nobits?", integer_is_nobits, sys::mrb_args_req(1))?
+        .add_method("size", integer_size, sys::mrb_args_none())?
         .define()?;
     interp.def_class::<Integer>(spec)?;
     let _ = interp.eval(&include_bytes!("integer.rb")[..])?;
@@ -26,7 +26,7 @@ pub fn init(interp: &mut Artichoke) -> InitializeResult<()> {
     Ok(())
 }
 
-unsafe extern "C" fn artichoke_integer_chr(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
+unsafe extern "C" fn integer_chr(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
     let encoding = mrb_get_args!(mrb, optional = 1);
     unwrap_interpreter!(mrb, to => guard);
     let value = Value::from(slf);
@@ -38,10 +38,7 @@ unsafe extern "C" fn artichoke_integer_chr(mrb: *mut sys::mrb_state, slf: sys::m
     }
 }
 
-unsafe extern "C" fn artichoke_integer_element_reference(
-    mrb: *mut sys::mrb_state,
-    slf: sys::mrb_value,
-) -> sys::mrb_value {
+unsafe extern "C" fn integer_element_reference(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
     let bit = mrb_get_args!(mrb, required = 1);
     unwrap_interpreter!(mrb, to => guard);
     let value = Value::from(slf);
@@ -53,7 +50,7 @@ unsafe extern "C" fn artichoke_integer_element_reference(
     }
 }
 
-unsafe extern "C" fn artichoke_integer_div(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
+unsafe extern "C" fn integer_div(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
     let denominator = mrb_get_args!(mrb, required = 1);
     unwrap_interpreter!(mrb, to => guard);
     let value = Value::from(slf);
@@ -65,7 +62,7 @@ unsafe extern "C" fn artichoke_integer_div(mrb: *mut sys::mrb_state, slf: sys::m
     }
 }
 
-unsafe extern "C" fn artichoke_integer_is_allbits(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
+unsafe extern "C" fn integer_is_allbits(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
     let mask = mrb_get_args!(mrb, required = 1);
     unwrap_interpreter!(mrb, to => guard);
     let value = Value::from(slf);
@@ -77,7 +74,7 @@ unsafe extern "C" fn artichoke_integer_is_allbits(mrb: *mut sys::mrb_state, slf:
     }
 }
 
-unsafe extern "C" fn artichoke_integer_is_anybits(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
+unsafe extern "C" fn integer_is_anybits(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
     let mask = mrb_get_args!(mrb, required = 1);
     unwrap_interpreter!(mrb, to => guard);
     let value = Value::from(slf);
@@ -89,7 +86,7 @@ unsafe extern "C" fn artichoke_integer_is_anybits(mrb: *mut sys::mrb_state, slf:
     }
 }
 
-unsafe extern "C" fn artichoke_integer_is_nobits(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
+unsafe extern "C" fn integer_is_nobits(mrb: *mut sys::mrb_state, slf: sys::mrb_value) -> sys::mrb_value {
     let mask = mrb_get_args!(mrb, required = 1);
     unwrap_interpreter!(mrb, to => guard);
     let value = Value::from(slf);
@@ -101,7 +98,7 @@ unsafe extern "C" fn artichoke_integer_is_nobits(mrb: *mut sys::mrb_state, slf: 
     }
 }
 
-unsafe extern "C" fn artichoke_integer_size(mrb: *mut sys::mrb_state, _slf: sys::mrb_value) -> sys::mrb_value {
+unsafe extern "C" fn integer_size(mrb: *mut sys::mrb_state, _slf: sys::mrb_value) -> sys::mrb_value {
     mrb_get_args!(mrb, none);
     unwrap_interpreter!(mrb, to => guard);
     let result = trampoline::size(&guard);
